Tried fixing how units move in group. Also minor arts changes. I've switched off diagonal walking for now, since you can't walk diagonally if the cardinal tiles are blocked anyway.

+++ src/simplerts/actions/Destination.java
@@ -11,9 +11,9 @@ package simplerts.actions;
  */
 public class Destination {
     
-    private float x, y;
+    private int x, y;
     
-    public Destination(float x, float y)
+    public Destination(int x, int y)
     {
         this.x = x;
         this.y = y;
@@ -22,28 +22,28 @@ public class Destination {
     /**
      * @return the x
      */
-    public float getX() {
+    public int getX() {
         return x;
     }
 
     /**
      * @param x the x to set
      */
-    public void setX(float x) {
+    public void setX(int x) {
         this.x = x;
     }
 
     /**
      * @return the y
      */
-    public float getY() {
+    public int getY() {
         return y;
     }
 
     /**
      * @param y the y to set
      */
-    public void setY(float y) {
+    public void setY(int y) {
         this.y = y;
     }
